# Remove many extra conversions from Matrix3x3 to Quaternion