--- a/tf2/include/tf2/buffer_core.h
+++ b/tf2/include/tf2/buffer_core.h
@@ -383,7 +383,7 @@ private:
 
   bool setTransformImpl(
     const tf2::Vector3 & origin_in, const tf2::Quaternion & rotation_in, const std::string & frame_id,
-    const std::string child_frame_id, const TimePoint stamp,
+    const std::string & child_frame_id, const TimePoint stamp,
     const std::string & authority, bool is_static);
   void lookupTransformImpl(
     const std::string & target_frame, const std::string & source_frame,