--- a/tf2/include/tf2/buffer_core.h
+++ b/tf2/include/tf2/buffer_core.h
@@ -382,15 +382,16 @@ private:
   std::string allFramesAsStringNoLock() const;
 
   bool setTransformImpl(
-    const tf2::Vector3 & origin_in, const tf2::Quaternion & rotation_in, const std::string & frame_id,
-    const std::string & child_frame_id, const TimePoint stamp,
+    const tf2::Vector3 & origin_in, const tf2::Quaternion & rotation_in,
+    const std::string & frame_id, const std::string & child_frame_id, const TimePoint stamp,
     const std::string & authority, bool is_static);
   void lookupTransformImpl(
     const std::string & target_frame, const std::string & source_frame,
     const TimePoint & time_in, tf2::Transform & transform_out, TimePoint & time_out) const;
   void lookupTransformImpl(
     const std::string & target_frame, const std::string & source_frame,
-    const TimePoint & time_in, tf2::Vector3 & origin_out, tf2::Quaternion & rotation_out, TimePoint & time_out) const;
+    const TimePoint & time_in, tf2::Vector3 & origin_out, tf2::Quaternion & rotation_out,
+    TimePoint & time_out) const;
 
   void lookupTransformImpl(
     const std::string & target_frame, const TimePoint & target_time,