--- a/tf2/src/buffer_core.cpp
+++ b/tf2/src/buffer_core.cpp
@@ -289,7 +289,7 @@ bool BufferCore::setTransformImpl(
 
     if (frame->insertData(
         TransformStorage(
-          stamp, transform_in.getRotation(),
+          stamp, rotation_in,
           transform_in.getOrigin(), lookupOrInsertFrameNumber(stripped_frame_id), frame_number)))
     {
       frame_authority_[frame_number] = authority;