--- a/tf2/src/buffer_core.cpp
+++ b/tf2/src/buffer_core.cpp
@@ -202,7 +202,7 @@ bool BufferCore::setTransform(
 
 bool BufferCore::setTransformImpl(
   const tf2::Vector3 & origin_in, const tf2::Quaternion & rotation_in, const std::string & frame_id,
-  const std::string child_frame_id, const TimePoint stamp,
+  const std::string & child_frame_id, const TimePoint stamp,
   const std::string & authority, bool is_static)
 {
   std::string stripped_frame_id = stripSlash(frame_id);