--- a/tf2/src/buffer_core.cpp
+++ b/tf2/src/buffer_core.cpp
@@ -737,7 +737,7 @@ BufferCore::lookupTransform(
   msg.transform.translation.x = transform.getOrigin().x();
   msg.transform.translation.y = transform.getOrigin().y();
   msg.transform.translation.z = transform.getOrigin().z();
-  tf2::Quaternion rotation = transform.getRotation();
+  const tf2::Quaternion rotation = transform.getRotation();
   msg.transform.rotation.x = rotation.x();
   msg.transform.rotation.y = rotation.y();
   msg.transform.rotation.z = rotation.z();