--- a/tf2/src/buffer_core.cpp
+++ b/tf2/src/buffer_core.cpp
@@ -183,14 +183,14 @@ bool BufferCore::setTransform(
   const std::string & authority, bool is_static)
 {
   tf2::Quaternion rotation(
-      transform.transform.rotation.x,
-      transform.transform.rotation.y,
-      transform.transform.rotation.z,
-      transform.transform.rotation.w);
+    transform.transform.rotation.x,
+    transform.transform.rotation.y,
+    transform.transform.rotation.z,
+    transform.transform.rotation.w);
   tf2::Vector3 origin(
-      transform.transform.translation.x,
-      transform.transform.translation.y,
-      transform.transform.translation.z);
+    transform.transform.translation.x,
+    transform.transform.translation.y,
+    transform.transform.translation.z);
   TimePoint time_point(std::chrono::nanoseconds(transform.header.stamp.nanosec) +
     std::chrono::duration_cast<std::chrono::nanoseconds>(
       std::chrono::seconds(
@@ -754,15 +754,15 @@ BufferCore::lookupTransform(
   return msg;
 }
 
- /** \brief Private member function that looks up a transform between two frames 
-    * at a given time and return the the transform as a TF2::transform. If a transform is 
+/** \brief Private member function that looks up a transform between two frames
+    * at a given time and return the the transform as a TF2::transform. If a transform is
     * not possible raise the appropriate error.
     * \param target_frame -- the name of the target frame which we are transforming to
     * \param source_frame -- the name of frame we are transforming from
     * \param time -- the timepoint at which the transform should occur
-    * \param transform_out -- The transform, returned by reference, for the transform from source frame 
+    * \param transform_out -- The transform, returned by reference, for the transform from source frame
     * to target frame at the given time
-    * \param time_out -- the time the transform was computed returned by reference.      
+    * \param time_out -- the time the transform was computed returned by reference.
     * \return void, the Transform between the input frames, and the time at which it was calculated are returned by reference
     */
 void BufferCore::lookupTransformImpl(
@@ -772,22 +772,23 @@ void BufferCore::lookupTransformImpl(
   TimePoint & time_out) const
 {
   tf2::Quaternion rotation;
-  lookupTransformImpl(target_frame, source_frame, time,
+  lookupTransformImpl(
+    target_frame, source_frame, time,
     transform_out.getOrigin(), rotation, time_out);
   transform_out.setRotation(rotation);
 }
 
- /** \brief Private member function that looks up a transform between two frames 
-    * at a given time and return the the transform as a TF2::transform. If a transform is 
+/** \brief Private member function that looks up a transform between two frames
+    * at a given time and return the the transform as a TF2::transform. If a transform is
     * not possible raise the appropriate error. The rotation and orientation
-    * components are returned separately by reference. 
+    * components are returned separately by reference.
     * \param target_frame -- the name of the target frame which we are transforming to
     * \param source_frame -- the name of frame we are transforming from
     * \param time -- the timepoint at which the transform should occur
     * \param origin_out -- the position component of the desired transform
     *  passed by reference
     * \param origin_out -- the position component of the desired transform
-    *  passed by reference    
+    *  passed by reference
     * \param rotation_out -- the rotation component of the desired transform
     *  passed by reference
     * \return void, the Transform between the input frames, and the time at which it was calculated are returned by reference