When calling `setTransform()`, rotation data would flow as  `ros_msgs::Quaternion` -> `tf2::Quaternion` -> `tf2::Matrix3x3`, then reconvert to `tf2::Quaternion` 12 times to check validity of the rotation, then do one final conversion to actually store it inside the buffer.

When calling `lookupTransform()`, we'd go from `tf2::Quaternion`->`tf2::Matrix3x3`->`tf2::Quaternion` (x4)->`ros_msgs::Quaternion`.

I've fixed this in two parts - 
1. Change `setTransformImpl` and `lookupTransformImpl` to take a vector+quat, eliminating both the roundtrip through `tf2::Transfrom` and the possible footgun of `getRotation`
2. In places that still need a `tf2::Transform`, only call `getRotation()` once, and reusing the result afterwards
3. I did not change the reference frame version of `lookupTransform` nor `lookupVelocity`, those are far less mechanical changes - would appreciate if someone fixed these up in a followup PR

Some notes:
1. `getRotation()` is poorly named - would recommend deprecating and renaming to `calculateRotation` or similar. I shot myself in the foot implementing `tf2_basis` and not realizing that origin is by reference and rotation is by value.
2. Part of the reason this happened is #732 - there are conversion functions that actually do the conversion correctly (not converting once per component) but they are unable to be used due to circular reference
3. It feels like `tf2::Transform` should really be a vector+quat, but it would be an API change