@kyle-basis 2024-12-04
Note: I've broken all tests, I'll fix it - the base idea is sound, however