@kyle-basis 2024-12-04
Pushed what should hopefully fix the failing tests. `q == -q`