@kyle-basis 2024-12-04
Hit a failed test in CI, I don't think this is my fault - ci flake?
```
02:29:00 5:  - /usr/bin/python3 -m launch_testing.launch_test /tmp/ws/src/geometry2/test_tf2/test/buffer_client_tester.launch.py --junit-xml=/tmp/ws/test_results/test_tf2/test_buffer_client_tester.launch.py.xunit.xml --package-name=test_tf2
02:29:00 5: [INFO] [launch]: All log files can be found below /home/buildfarm/.ros/log/2024-12-04-02-29-00-675074-4d63d25031d7-4049
02:29:00 5: [INFO] [launch]: Default logging verbosity is set to INFO
02:29:00 5: test_termination (test_tf2.TestBufferClient.test_termination) ... [INFO] [static_transform_publisher-1]: process started with pid [4053]
02:29:00 5: [INFO] [test_buffer_server-2]: process started with pid [4054]
02:29:00 5: [INFO] [test_buffer_client-3]: process started with pid [4055]
02:29:00 5: [INFO] [python3-4]: process started with pid [4056]
02:29:00 5: [static_transform_publisher-1] [INFO] [1733308140.776567271] [static_transform_publisher_VNHQnJSRyXBSDNpK]: Spinning until stopped - publishing transform
02:29:00 5: [static_transform_publisher-1] translation: ('5.000000', '6.000000', '7.000000')
02:29:00 5: [static_transform_publisher-1] rotation: ('0.000000', '0.000000', '0.000000', '1.000000')
02:29:00 5: [static_transform_publisher-1] from 'a' to 'b'
02:29:00 5: [test_buffer_client-3] [==========] Running 2 tests from 1 test suite.
02:29:01 5: [test_buffer_client-3] [----------] Global test environment set-up.
02:29:01 5: [test_buffer_client-3] [----------] 2 tests from tf2_ros
02:29:01 5: [test_buffer_client-3] [ RUN      ] tf2_ros.buffer_client
02:29:01 5: [test_buffer_client-3] FOOO
02:29:01 5: [test_buffer_client-3] [INFO] [1733308142.718698560] [tf_action_node]: p1: (0.00, 0.00, 0.00), p2: (-5.00, -6.00, -7.00)
02:29:02 5: [test_buffer_client-3] [       OK ] tf2_ros.buffer_client (1014 ms)
02:29:02 5: [test_buffer_client-3] [ RUN      ] tf2_ros.buffer_client_different_types
02:29:02 5: [test_buffer_client-3] [ERROR] [1733308142.795791742] [tf_action_node]: Bullet: (-1.0000, 0.0000, 0.0000)
02:29:02 5: [test_buffer_client-3] [ERROR] [1733308142.795866165] [tf_action_node]: KDL: (0.0000, 0.0000, 0.0000)
02:29:02 5: [test_buffer_client-3] /tmp/ws/src/geometry2/test_tf2/test/test_buffer_client.cpp:127: Failure
02:29:02 5: [test_buffer_client-3] The difference between b1[0] and -5.0 is 4, which exceeds EPS, where
02:29:02 5: [test_buffer_client-3] b1[0] evaluates to -1,
02:29:02 5: [test_buffer_client-3] -5.0 evaluates to -5, and
02:29:02 5: [test_buffer_client-3] EPS evaluates to 0.001.
02:29:02 5: [test_buffer_client-3] 
02:29:02 5: [test_buffer_client-3] /tmp/ws/src/geometry2/test_tf2/test/test_buffer_client.cpp:128: Failure
02:29:02 5: [test_buffer_client-3] The difference between b1[1] and -6.0 is 6, which exceeds EPS, where
02:29:02 5: [test_buffer_client-3] b1[1] evaluates to 0,
02:29:02 5: [test_buffer_client-3] -6.0 evaluates to -6, and
02:29:02 5: [test_buffer_client-3] EPS evaluates to 0.001.
02:29:02 5: [test_buffer_client-3] 
02:29:02 5: [test_buffer_client-3] /tmp/ws/src/geometry2/test_tf2/test/test_buffer_client.cpp:129: Failure
02:29:02 5: [test_buffer_client-3] The difference between b1[2] and -7.0 is 7, which exceeds EPS, where
02:29:02 5: [test_buffer_client-3] b1[2] evaluates to 0,
02:29:02 5: [test_buffer_client-3] -7.0 evaluates to -7, and
02:29:02 5: [test_buffer_client-3] EPS evaluates to 0.001.
02:29:02 5: [test_buffer_client-3] 
02:29:02 5: [test_buffer_client-3] [  FAILED  ] tf2_ros.buffer_client_different_types (74 ms)
02:29:02 5: [test_buffer_client-3] [----------] 2 tests from tf2_ros (1089 ms total)
02:29:02 5: [test_buffer_client-3] 
02:29:02 5: [test_buffer_client-3] [----------] Global test environment tear-down
02:29:02 5: [test_buffer_client-3] [==========] 2 tests from 1 test suite ran. (1089 ms total)
02:29:02 5: [test_buffer_client-3] [  PASSED  ] 1 test.
02:29:02 5: [test_buffer_client-3] [  FAILED  ] 1 test, listed below:
02:29:02 5: [test_buffer_client-3] [  FAILED  ] tf2_ros.buffer_client_different_types
02:29:02 5: [test_buffer_client-3] 
02:29:02 5: [test_buffer_client-3]  1 FAILED TEST
```

Locally:
```
[test_buffer_client-3] [==========] Running 2 tests from 1 test suite.
[test_buffer_client-3] [----------] Global test environment set-up.
[test_buffer_client-3] [----------] 2 tests from tf2_ros
[test_buffer_client-3] [ RUN      ] tf2_ros.buffer_client
[test_buffer_client-3] FOOO
[test_buffer_client-3] [INFO] [1733348400.995224948] [tf_action_node]: p1: (0.00, 0.00, 0.00), p2: (-5.00, -6.00, -7.00)
[test_buffer_client-3] [       OK ] tf2_ros.buffer_client (403 ms)
[test_buffer_client-3] [ RUN      ] tf2_ros.buffer_client_different_types
[test_buffer_client-3] [ERROR] [1733348401.032240829] [tf_action_node]: Bullet: (-5.0000, -6.0000, -7.0000)
[test_buffer_client-3] [ERROR] [1733348401.032402204] [tf_action_node]: KDL: (0.0000, 0.0000, 0.0000)
[test_buffer_client-3] [       OK ] tf2_ros.buffer_client_different_types (30 ms)
[test_buffer_client-3] [----------] 2 tests from tf2_ros (434 ms total)
[test_buffer_client-3] 
[test_buffer_client-3] [----------] Global test environment tear-down
[test_buffer_client-3] [==========] 2 tests from 1 test suite ran. (434 ms total)
[test_buffer_client-3] [  PASSED  ] 2 tests.
[INFO] [test_buffer_client-3]: process has finished cleanly [pid 254386]
ok
```