@kscottz 2024-12-04
Sorry about the delay, had a power outage this morning.  I'll take a look but I am not super familiar with the internals of TF / Geometry. Thanks for sticking with this and trying to get it over the line. We really appreciate the help and could always use a few more contributors. ❤️ 

> Note: I've broken all tests, I'll fix it - the base idea is sound, however

I figured you might hit a few flakey tests as anything that relies on float / double ends up being a bit brittle. I can't speak to   any of these tests being historically flakey. I think setting the epsilons to five or six sig figs should be sufficient. 

> This could probably use some documentation

Absolutely. We're starting to have conversations about bringing in someone to address API doc deficiencies but that's still far off. I would need more familiarity to do it myself. 

> Never mind - appears to just be tests using hardcoded quaternion values, rather than actually checking equality - I'm happy to fix this either by manually fixing the test quaternions, or by doing some nicer equality check. This does mean that downstream users will sometimes get out different (but equiv) rotations from the library than they got before.

This one's on you and what you're up for. I would have to look at the individual tests to make a call. Please do keep in mind that more you change on the lower level code the more likely it is that it may break something up stream. Smaller deltas are always better and get finished faster. You can always circle back

> isNan() or isValid() on Vector/Quaternion would be really nice for this, along with a normalization check on quaternion - this code doesn't belong here.


I'll need to look at these a bit more but probably warrants an issue with a `good first issue` tag.