@kyle-basis 2024-12-27
@kscottz wrote a benchmark:
```cpp


#include <benchmark/benchmark.h>

#include "tf2_ros/buffer.h"

static void benchmark_set_transform(benchmark::State & state)
{
    rclcpp::Clock::SharedPtr clock = std::make_shared<rclcpp::Clock>(RCL_SYSTEM_TIME);
    tf2_ros::Buffer buffer(clock);
    buffer.setUsingDedicatedThread(false);

    geometry_msgs::msg::TransformStamped transform;
    transform.header.frame_id = "foo";
    transform.header.stamp.sec = 0;
    transform.header.stamp.nanosec = 0;
    transform.child_frame_id = "bar";
    transform.transform.translation.x = 42.0;
    transform.transform.translation.y = -3.14;
    transform.transform.translation.z = 0.0;
    transform.transform.rotation.x = 0.591922;
    transform.transform.rotation.y = 0.210684;
    transform.transform.rotation.z = -0.762476;
    transform.transform.rotation.w = 0.154502;

    for (auto _ : state) {
        transform.header.stamp.nanosec++;
        buffer.setTransform(transform, "benchmark");   
    }
}

BENCHMARK(benchmark_set_transform);

before:
```
basis@ubuntu:~/tf2_ws/src/geometry2$ ./build/tf2_ros/buffer_benchmark --benchmark_repetitions=10
2024-12-27T14:54:09-08:00
Running ./build/tf2_ros/buffer_benchmark
Run on (8 X 2201.6 MHz CPU s)
CPU Caches:
  L1 Data 64 KiB (x8)
  L1 Instruction 64 KiB (x8)
  L2 Unified 256 KiB (x8)
  L3 Unified 2048 KiB (x2)
Load Average: 0.57, 1.27, 1.16
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
-------------------------------------------------------------------------
Benchmark                               Time             CPU   Iterations
-------------------------------------------------------------------------
benchmark_set_transform             80170 ns        80080 ns        23146
benchmark_set_transform             94480 ns        94391 ns        23146
benchmark_set_transform             64783 ns        64723 ns        23146
benchmark_set_transform             64649 ns        64590 ns        23146
benchmark_set_transform             85981 ns        85908 ns        23146
benchmark_set_transform             64721 ns        64663 ns        23146
benchmark_set_transform             93638 ns        93547 ns        23146
benchmark_set_transform             99861 ns        99763 ns        23146
benchmark_set_transform            103710 ns       103613 ns        23146
benchmark_set_transform             64942 ns        64883 ns        23146
benchmark_set_transform_mean        81694 ns        81616 ns           10
benchmark_set_transform_median      83076 ns        82994 ns           10
benchmark_set_transform_stddev      15940 ns        15924 ns           10
benchmark_set_transform_cv          19.51 %         19.51 %            10
```

after
```
basis@ubuntu:~/tf2_ws/src/geometry2$ ./build/tf2_ros/buffer_benchmark --benchmark_repetitions=10
2024-12-27T14:58:55-08:00
Running ./build/tf2_ros/buffer_benchmark
Run on (8 X 2201.6 MHz CPU s)
CPU Caches:
  L1 Data 64 KiB (x8)
  L1 Instruction 64 KiB (x8)
  L2 Unified 256 KiB (x8)
  L3 Unified 2048 KiB (x2)
Load Average: 2.31, 1.35, 1.16
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
-------------------------------------------------------------------------
Benchmark                               Time             CPU   Iterations
-------------------------------------------------------------------------
benchmark_set_transform             79914 ns        79811 ns        23001
benchmark_set_transform             94020 ns        93897 ns        23001
benchmark_set_transform             64227 ns        64148 ns        23001
benchmark_set_transform             64270 ns        64194 ns        23001
benchmark_set_transform             88543 ns        88430 ns        23001
benchmark_set_transform             64285 ns        64207 ns        23001
benchmark_set_transform             84061 ns        83955 ns        23001
benchmark_set_transform            101408 ns       101276 ns        23001
benchmark_set_transform            106780 ns       106642 ns        23001
benchmark_set_transform             64476 ns        64398 ns        23001
benchmark_set_transform_mean        81198 ns        81096 ns           10
benchmark_set_transform_median      81988 ns        81883 ns           10
benchmark_set_transform_stddev      16436 ns        16412 ns           10
benchmark_set_transform_cv          20.24 %         20.24 %            10
```

This was run on a Jetson Orin kit, with cpu scaling governer set to `performance`. TBH the results aren't amazing - it looks to be around 2% faster (and sometimes slightly slower - struggling to actually disable CPU scaling on this machine). I'd suspected this to be the case - if it were an actual performance problem, someone would have raised a flag before me. I think in terms of code quality this is the correct solution, however.